Updates for DB Support

+++ src/main/java/com/bah/project/dataservice/repository/RegistrationRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import com.bah.project.dataservice.domain.Registration;
 
 @Repository
-public interface RegistrationRepository extends CrudRepository<Registration, Long>{
+public interface RegistrationRepository extends CrudRepository<Registration, Integer>{
 
 	
 }
